// src/shared/types/cv-graphql.ts
export enum Mastery {
  Novice = "Novice",
  Advanced = "Advanced",
  Competent = "Competent",
  Proficient = "Proficient",
  Expert = "Expert",
}
// ...
export interface SkillMastery {
  name: string;
  categoryId?: Nullable<string>;
  mastery: Mastery;
}
// ...
type Nullable<T> = T | null;